+++ crates/zng-wgt-shortcut/src/lib.rs
@@ -49,28 +49,31 @@ context_var! {
     pub static FIRST_N_VAR: usize = 1;
 
     /// Widget function that generates the outer panel.
-    pub static PANEL_FN_VAR: WidgetFn<PanelArgs> = WidgetFn::new(default_panel_fn);
+    pub static PANEL_FN_VAR: WidgetFn<PanelFnArgs> = WidgetFn::new(default_panel_fn);
 
     /// Widget function that generates the separator between shortcuts.
-    pub static SHORTCUTS_SEPARATOR_FN_VAR: WidgetFn<ShortcutsSeparatorArgs> = WidgetFn::new(default_shortcuts_separator_fn);
+    pub static SHORTCUTS_SEPARATOR_FN_VAR: WidgetFn<ShortcutsSeparatorFnArgs> = WidgetFn::new(default_shortcuts_separator_fn);
 
     /// Widget function that generates a shortcut panel.
-    pub static SHORTCUT_FN_VAR: WidgetFn<ShortcutArgs> = WidgetFn::nil();
+    pub static SHORTCUT_FN_VAR: WidgetFn<ShortcutFnArgs> = WidgetFn::nil();
 
     /// Widget function that generates the separator between key gestures in chord shortcuts.
-    pub static CHORD_SEPARATOR_FN_VAR: WidgetFn<ChordSeparatorArgs> = WidgetFn::new(default_chord_separator_fn);
+    pub static CHORD_SEPARATOR_FN_VAR: WidgetFn<ChordSeparatorFnArgs> = WidgetFn::new(default_chord_separator_fn);
 
     /// Widget function that generates the modifier view.
-    pub static MODIFIER_FN_VAR: WidgetFn<ModifierArgs> = WidgetFn::new(default_modifier_fn);
+    pub static MODIFIER_FN_VAR: WidgetFn<ModifierFnArgs> = WidgetFn::new(default_modifier_fn);
 
     /// Widget function that generates the key gesture panel.
-    pub static KEY_GESTURE_FN_VAR: WidgetFn<KeyGestureArgs> = WidgetFn::nil();
+    pub static KEY_GESTURE_FN_VAR: WidgetFn<KeyGestureFnArgs> = WidgetFn::nil();
 
     /// Widget function that generates the separators between modifiers and keys in a key gesture.
-    pub static KEY_GESTURE_SEPARATOR_FN_VAR: WidgetFn<KeyGestureSeparatorArgs> = WidgetFn::new(default_key_gesture_separator_fn);
+    pub static KEY_GESTURE_SEPARATOR_FN_VAR: WidgetFn<KeyGestureSeparatorFnArgs> = WidgetFn::new(default_key_gesture_separator_fn);
 
     /// Widget function that generates the key view.
-    pub static KEY_FN_VAR: WidgetFn<KeyArgs> = WidgetFn::new(default_key_fn);
+    pub static KEY_FN_VAR: WidgetFn<KeyFnArgs> = WidgetFn::new(default_key_fn);
+
+    /// Widget function that generates content when there is no gesture to display.
+    pub static NONE_FN_VAR: WidgetFn<NoneFnArgs> = WidgetFn::nil();
 }
 
 /// Maximum number of shortcuts to display.
@@ -81,15 +84,23 @@ pub fn first_n(child: impl IntoUiNode, n: impl IntoVar<usize>) -> UiNode {
     with_context_var(child, FIRST_N_VAR, n)
 }
 
-/// Widget function that converts [`PanelArgs`] to a widget.
+/// Widget function that converts [`PanelFnArgs`] to a widget.
 ///
 /// This property sets the [`PANEL_FN_VAR`].
 #[property(CONTEXT, default(PANEL_FN_VAR), widget_impl(ShortcutText))]
-pub fn panel_fn(child: impl IntoUiNode, panel_fn: impl IntoVar<WidgetFn<PanelArgs>>) -> UiNode {
+pub fn panel_fn(child: impl IntoUiNode, panel_fn: impl IntoVar<WidgetFn<PanelFnArgs>>) -> UiNode {
     with_context_var(child, PANEL_FN_VAR, panel_fn)
 }
 
-/// Widget function that converts [`ShortcutsSeparatorArgs`] to a widget.
+/// Widget function that converts [`NoneFnArgs`] to a widget.
+///
+/// This property sets the [`NONE_FN_VAR`].
+#[property(CONTEXT, default(NONE_FN_VAR), widget_impl(ShortcutText))]
+pub fn none_fn(child: impl IntoUiNode, none_fn: impl IntoVar<WidgetFn<NoneFnArgs>>) -> UiNode {
+    with_context_var(child, NONE_FN_VAR, none_fn)
+}
+
+/// Widget function that converts [`ShortcutsSeparatorFnArgs`] to a widget.
 ///
 /// This is the separators between shortcuts, when [`first_n`] is more than one and the [`shortcut`] has mode them one shortcut.
 ///
@@ -100,11 +111,11 @@ pub fn panel_fn(child: impl IntoUiNode, panel_fn: impl IntoVar<WidgetFn<PanelArg
 /// [`first_n`]: fn@first_n
 /// [`shortcut`]: fn@shortcut
 #[property(CONTEXT, default(SHORTCUTS_SEPARATOR_FN_VAR), widget_impl(ShortcutText))]
-pub fn shortcuts_separator_fn(child: impl IntoUiNode, separator_fn: impl IntoVar<WidgetFn<ShortcutsSeparatorArgs>>) -> UiNode {
+pub fn shortcuts_separator_fn(child: impl IntoUiNode, separator_fn: impl IntoVar<WidgetFn<ShortcutsSeparatorFnArgs>>) -> UiNode {
     with_context_var(child, SHORTCUTS_SEPARATOR_FN_VAR, separator_fn)
 }
 
-/// Widget function that converts [`ShortcutArgs`] to a widget.
+/// Widget function that converts [`ShortcutFnArgs`] to a widget.
 ///
 /// Set to [`WidgetFn::nil`] or return the items as a node to pass the items directly to [`panel_fn`].
 ///
@@ -112,21 +123,21 @@ pub fn shortcuts_separator_fn(child: impl IntoUiNode, separator_fn: impl IntoVar
 ///
 /// [`panel_fn`]: fn@panel_fn
 #[property(CONTEXT, default(SHORTCUT_FN_VAR), widget_impl(ShortcutText))]
-pub fn shortcut_fn(child: impl IntoUiNode, panel_fn: impl IntoVar<WidgetFn<ShortcutArgs>>) -> UiNode {
+pub fn shortcut_fn(child: impl IntoUiNode, panel_fn: impl IntoVar<WidgetFn<ShortcutFnArgs>>) -> UiNode {
     with_context_var(child, SHORTCUT_FN_VAR, panel_fn)
 }
 
-/// Widget function that converts [`ChordSeparatorArgs`] to a widget.
+/// Widget function that converts [`ChordSeparatorFnArgs`] to a widget.
 ///
 /// This is the separator between the starter and complement in a [`KeyChord`].
 ///
 /// This property sets the [`CHORD_SEPARATOR_FN_VAR`].
 #[property(CONTEXT, default(CHORD_SEPARATOR_FN_VAR), widget_impl(ShortcutText))]
-pub fn chord_separator_fn(child: impl IntoUiNode, separator_fn: impl IntoVar<WidgetFn<ChordSeparatorArgs>>) -> UiNode {
+pub fn chord_separator_fn(child: impl IntoUiNode, separator_fn: impl IntoVar<WidgetFn<ChordSeparatorFnArgs>>) -> UiNode {
     with_context_var(child, CHORD_SEPARATOR_FN_VAR, separator_fn)
 }
 
-/// Widget function that converts [`KeyGestureArgs`] to a widget.
+/// Widget function that converts [`KeyGestureFnArgs`] to a widget.
 ///
 /// Set to [`WidgetFn::nil`] or return the items as a node to pass the items directly to [`shortcut_fn`].
 ///
@@ -134,35 +145,35 @@ pub fn chord_separator_fn(child: impl IntoUiNode, separator_fn: impl IntoVar<Wid
 ///
 /// [`shortcut_fn`]: fn@shortcut_fn
 #[property(CONTEXT, default(KEY_GESTURE_FN_VAR), widget_impl(ShortcutText))]
-pub fn key_gesture_fn(child: impl IntoUiNode, panel_fn: impl IntoVar<WidgetFn<KeyGestureArgs>>) -> UiNode {
+pub fn key_gesture_fn(child: impl IntoUiNode, panel_fn: impl IntoVar<WidgetFn<KeyGestureFnArgs>>) -> UiNode {
     with_context_var(child, KEY_GESTURE_FN_VAR, panel_fn)
 }
 
-/// Widget function that converts [`KeyGestureSeparatorArgs`] to a widget.
+/// Widget function that converts [`KeyGestureSeparatorFnArgs`] to a widget.
 ///
 /// This is the separator between the modifiers and key in a [`KeyGesture`].
 ///
 /// This property sets the [`KEY_GESTURE_SEPARATOR_FN_VAR`].
 #[property(CONTEXT, default(KEY_GESTURE_SEPARATOR_FN_VAR), widget_impl(ShortcutText))]
-pub fn key_gesture_separator_fn(child: impl IntoUiNode, separator_fn: impl IntoVar<WidgetFn<KeyGestureSeparatorArgs>>) -> UiNode {
+pub fn key_gesture_separator_fn(child: impl IntoUiNode, separator_fn: impl IntoVar<WidgetFn<KeyGestureSeparatorFnArgs>>) -> UiNode {
     with_context_var(child, KEY_GESTURE_SEPARATOR_FN_VAR, separator_fn)
 }
 
-/// Widget function that converts a [`ModifierArgs`] to a widget.
+/// Widget function that converts a [`ModifierFnArgs`] to a widget.
 ///
 /// This is used for both the [`Shortcut::Modifier`] standalone and the [`KeyGesture::modifiers`] flags.
 ///
 /// This property sets the [`MODIFIER_FN_VAR`].
 #[property(CONTEXT, default(MODIFIER_FN_VAR), widget_impl(ShortcutText))]
-pub fn modifier_fn(child: impl IntoUiNode, modifier_fn: impl IntoVar<WidgetFn<ModifierArgs>>) -> UiNode {
+pub fn modifier_fn(child: impl IntoUiNode, modifier_fn: impl IntoVar<WidgetFn<ModifierFnArgs>>) -> UiNode {
     with_context_var(child, MODIFIER_FN_VAR, modifier_fn)
 }
 
-/// Widget function that converts a [`KeyArgs`] to a widget.
+/// Widget function that converts a [`KeyFnArgs`] to a widget.
 ///  
 /// This property sets the [`KEY_FN_VAR`].
 #[property(CONTEXT, default(KEY_FN_VAR), widget_impl(ShortcutText))]
-pub fn key_fn(child: impl IntoUiNode, key_fn: impl IntoVar<WidgetFn<KeyArgs>>) -> UiNode {
+pub fn key_fn(child: impl IntoUiNode, key_fn: impl IntoVar<WidgetFn<KeyFnArgs>>) -> UiNode {
     with_context_var(child, KEY_FN_VAR, key_fn)
 }
 
@@ -170,25 +181,36 @@ pub fn key_fn(child: impl IntoUiNode, key_fn: impl IntoVar<WidgetFn<KeyArgs>>) -
 ///
 /// [`panel_fn`]: fn@panel_fn
 #[non_exhaustive]
-pub struct PanelArgs {
+pub struct PanelFnArgs {
     /// Shortcut and shortcut separator items.
     pub items: UiVec,
+
+    /// If the single item in `items` is the [`none_fn`].
+    ///
+    /// [`none_fn`]: fn@none_fn
+    pub is_none: bool,
 
     /// The shortcuts that where used to generate the `items`.
     pub shortcuts: Shortcuts,
 }
 
+/// Arguments for [`none_fn`].
+///
+/// [`none_fn`]: fn@none_fn
+#[non_exhaustive]
+pub struct NoneFnArgs {}
+
 /// Arguments for [`shortcuts_separator_fn`].
 ///
 /// [`shortcuts_separator_fn`]: fn@shortcuts_separator_fn
 #[non_exhaustive]
-pub struct ShortcutsSeparatorArgs {}
+pub struct ShortcutsSeparatorFnArgs {}
 
 /// Arguments for [`shortcut_fn`].
 ///
 /// [`shortcut_fn`]: fn@shortcut_fn
 #[non_exhaustive]
-pub struct ShortcutArgs {
+pub struct ShortcutFnArgs {
     /// Modifier, key and separator items.
     pub items: UiVec,
     /// The shortcut.
@@ -201,13 +223,13 @@ pub struct ShortcutArgs {
 ///
 /// [`chord_separator_fn`]: fn@chord_separator_fn
 #[non_exhaustive]
-pub struct ChordSeparatorArgs {}
+pub struct ChordSeparatorFnArgs {}
 
 /// Arguments for [`key_gesture_fn`].
 ///
 /// [`key_gesture_fn`]: fn@key_gesture_fn
 #[non_exhaustive]
-pub struct KeyGestureArgs {
+pub struct KeyGestureFnArgs {
     /// Modifier, key and separator items.
     pub items: UiVec,
     /// The key gesture.
@@ -220,19 +242,21 @@ pub struct KeyGestureArgs {
 ///
 /// [`modifier_fn`]: fn@modifier_fn
 #[non_exhaustive]
-pub struct ModifierArgs {
-    /// The modifier gesture.
-    pub gesture: ModifierGesture,
-    /// If is actually the [`Shortcut::Modifier`] press and release modifier gesture.
+pub struct ModifierFnArgs {
+    /// The modifier.
+    pub modifier: ModifierGesture,
+    /// If is actually the [`Shortcut::Modifier`] press and release gesture.
     ///
     /// If `false` is actually a [`ModifiersState`] flag extracted from [`KeyGesture::modifiers`].
+    ///
+    /// [`ModifiersState`]: zng_app::shortcut::ModifiersState
     pub is_standalone: bool,
 }
 
 /// Arguments for [`key_fn`].
 ///
 /// [`key_fn`]: fn@key_fn
-pub struct KeyArgs {
+pub struct KeyFnArgs {
     /// The key.
     pub key: GestureKey,
 }
@@ -241,7 +265,7 @@ pub struct KeyArgs {
 ///
 /// [`key_gesture_separator_fn`]: fn@key_gesture_separator_fn
 #[non_exhaustive]
-pub struct KeyGestureSeparatorArgs {
+pub struct KeyGestureSeparatorFnArgs {
     /// If the separator will be placed between two modifiers.
     ///
     /// When this is `false` the separator is placed between a modifier and the key.
@@ -251,7 +275,7 @@ pub struct KeyGestureSeparatorArgs {
 /// Default value for [`PANEL_FN_VAR`].
 ///
 /// For zero items returns nil, for one item just returns the item, for more returns a `Wrap!`.
-pub fn default_panel_fn(mut args: PanelArgs) -> UiNode {
+pub fn default_panel_fn(mut args: PanelFnArgs) -> UiNode {
     match args.items.len() {
         0 => UiNode::nil(),
         1 => args.items.remove(0),
@@ -262,27 +286,27 @@ pub fn default_panel_fn(mut args: PanelArgs) -> UiNode {
 /// Default value for [`SHORTCUTS_SEPARATOR_VAR`].
 ///
 /// Returns `Text!(" or ")`.
-pub fn default_shortcuts_separator_fn(_: ShortcutsSeparatorArgs) -> UiNode {
+pub fn default_shortcuts_separator_fn(_: ShortcutsSeparatorFnArgs) -> UiNode {
     zng_wgt_text::Text!(" or ")
 }
 
 /// Default value for [`CHORD_SEPARATOR_VAR`].
 ///
 /// Returns `Text!(", ")`.
-pub fn default_chord_separator_fn(_: ChordSeparatorArgs) -> UiNode {
+pub fn default_chord_separator_fn(_: ChordSeparatorFnArgs) -> UiNode {
     zng_wgt_text::Text!(", ")
 }
 
 /// Default value for [`KEY_GESTURE_SEPARATOR_FN_VAR`].
 ///
 /// Returns `Text!("+")`.
-pub fn default_key_gesture_separator_fn(_: KeyGestureSeparatorArgs) -> UiNode {
+pub fn default_key_gesture_separator_fn(_: KeyGestureSeparatorFnArgs) -> UiNode {
     zng_wgt_text::Text!("+")
 }
 
-/// Default value for [`MODIFIER_FN_VAR`].
-pub fn default_modifier_fn(args: ModifierArgs) -> UiNode {
-    let txt = match args.gesture {
+/// Gets the localized modifier name.
+pub fn modifier_txt(modifier: ModifierGesture) -> Var<Txt> {
+    match modifier {
         // !!: TODO localize
         ModifierGesture::Super => {
             if cfg!(windows) {
@@ -297,20 +321,13 @@ pub fn default_modifier_fn(args: ModifierArgs) -> UiNode {
         ModifierGesture::Shift => "⇧Shift",
         ModifierGesture::Alt => "Alt",
     }
-    .to_txt();
-    key_text(
-        txt,
-        if args.is_standalone {
-            BorderStyle::Double
-        } else {
-            BorderStyle::Solid
-        },
-    )
+    .to_txt()
+    .into_var()
 }
 
-/// Default value for [`KEY_FN_VAR`].
-pub fn default_key_fn(args: KeyArgs) -> UiNode {
-    let txt = match args.key {
+/// Gets the localized key name.
+pub fn key_txt(key: GestureKey) -> Var<Txt> {
+    match key {
         GestureKey::Key(key) => match key {
             Key::Char(c) => c.to_uppercase().to_txt(),
             Key::Str(s) => s,
@@ -322,30 +339,54 @@ pub fn default_key_fn(args: KeyArgs) -> UiNode {
             key => key.name().to_txt(), // !!: TODO localize name
         },
         GestureKey::Code(key_code) => formatx!("{key_code:?}"),
-    };
-    key_text(txt, BorderStyle::Solid)
+    }
+    .into_var()
 }
 
-fn key_text(txt: Txt, border_style: BorderStyle) -> UiNode {
-    fn dark_or_light(pct: FactorPercent) -> Var<Rgba> {
-        expr_var! {
-            let base = *#{zng_wgt_text::FONT_COLOR_VAR};
-            match #{zng_color::COLOR_SCHEME_VAR} {
-                ColorScheme::Dark => colors::BLACK.with_alpha(pct).mix_normal(base),
-                ColorScheme::Light => colors::WHITE.with_alpha(pct).mix_normal(base),
-                _ => zng_color::colors::BLACK.with_alpha(100.pct()),
-            }
-        }
-    }
+/// Default value for [`MODIFIER_FN_VAR`].
+///
+/// Returns a [`default_keycap`] with the [`modifier_txt`].
+pub fn default_modifier_fn(args: ModifierFnArgs) -> UiNode {
+    keycap(modifier_txt(args.modifier), args.is_standalone)
+}
+
+/// Default value for [`KEY_FN_VAR`].
+///
+/// Returns a [`default_keycap`] with the [`key_txt`].
+pub fn default_key_fn(args: KeyFnArgs) -> UiNode {
+    keycap(key_txt(args.key), false)
+}
+
+/// Widget used b the [`default_modifier_fn`] and [`default_key_fn`] to render a `Text!` styled to look like a keycap.
+pub fn keycap(txt: Var<Txt>, is_standalone_modifier: bool) -> UiNode {
     zng_wgt_text::Text! {
         txt;
         font_family = ["Consolas", "Lucida Console", "monospace"];
         zng_wgt::border = {
-            widths: (1.dip(), 0.06.em().max(1.dip()), 0.06.em().max(1.dip()), 1.dip()),
-            sides: dark_or_light(80.pct()).map(move |&c| BorderSides::new_all((c, border_style))),
+            widths: if is_standalone_modifier { 0.2 } else { 0.08 }.em().max(1.dip()),
+            sides: expr_var! {
+                let base = *#{zng_wgt_text::FONT_COLOR_VAR};
+                let color = match #{zng_color::COLOR_SCHEME_VAR} {
+                    ColorScheme::Dark => colors::BLACK.with_alpha(70.pct()).mix_normal(base),
+                    ColorScheme::Light => colors::WHITE.with_alpha(70.pct()).mix_normal(base),
+                    _ => base.with_alpha(30.pct()),
+                };
+                BorderSides::new_all((
+                    color,
+                    if is_standalone_modifier {
+                        BorderStyle::Double
+                    } else {
+                        BorderStyle::Solid
+                    },
+                ))
+            },
         };
-        zng_wgt_fill::background_color = dark_or_light(100.pct());
-        zng_wgt::corner_radius = 0.1.em();
+        zng_wgt_fill::background_color = zng_color::COLOR_SCHEME_VAR.map(|c| match c {
+            ColorScheme::Dark => colors::BLACK,
+            ColorScheme::Light => colors::WHITE,
+            _ => zng_color::colors::BLACK.with_alpha(100.pct()),
+        });
+        zng_wgt::corner_radius = 0.2.em();
         txt_align = Align::START;
         zng_wgt::align = Align::START;
         zng_wgt_container::padding = (0, 0.20.em(), -0.15.em(), 0.20.em());
@@ -366,7 +407,8 @@ fn node(shortcut: Var<Shortcuts>) -> UiNode {
                 .sub_var(&KEY_FN_VAR)
                 .sub_var(&KEY_GESTURE_SEPARATOR_FN_VAR)
                 .sub_var(&KEY_GESTURE_FN_VAR)
-                .sub_var(&FIRST_N_VAR);
+                .sub_var(&FIRST_N_VAR)
+                .sub_var(&NONE_FN_VAR);
             *c.node() = generate(shortcut.get());
             c.init();
         }
@@ -385,6 +427,7 @@ fn node(shortcut: Var<Shortcuts>) -> UiNode {
                 || KEY_GESTURE_SEPARATOR_FN_VAR.is_new()
                 || KEY_GESTURE_FN_VAR.is_new()
                 || FIRST_N_VAR.is_new()
+                || NONE_FN_VAR.is_new()
             {
                 c.deinit();
                 *c.node() = generate(shortcut.get());
@@ -397,7 +440,7 @@ fn node(shortcut: Var<Shortcuts>) -> UiNode {
     })
 }
 
-fn generate(shortcut: Shortcuts) -> UiNode {
+fn generate(mut shortcut: Shortcuts) -> UiNode {
     let panel_fn = PANEL_FN_VAR.get();
     let shortcuts_separator_fn = SHORTCUTS_SEPARATOR_FN_VAR.get();
     let shortcut_fn = SHORTCUT_FN_VAR.get();
@@ -408,10 +451,12 @@ fn generate(shortcut: Shortcuts) -> UiNode {
     let key_fn = KEY_FN_VAR.get();
     let first_n = FIRST_N_VAR.get();
 
+    shortcut.truncate(first_n);
+
     let mut items = ui_vec![];
-    for shortcut in shortcut.iter().take(first_n) {
+    for shortcut in shortcut.iter() {
         if !items.is_empty()
-            && let Some(sep) = shortcuts_separator_fn.call_checked(ShortcutsSeparatorArgs {})
+            && let Some(sep) = shortcuts_separator_fn.call_checked(ShortcutsSeparatorFnArgs {})
         {
             items.push(sep);
         }
@@ -419,23 +464,23 @@ fn generate(shortcut: Shortcuts) -> UiNode {
         fn gesture(
             out: &mut UiVec,
             gesture: KeyGesture,
-            separator_fn: &WidgetFn<KeyGestureSeparatorArgs>,
-            modifier_fn: &WidgetFn<ModifierArgs>,
-            key_fn: &WidgetFn<KeyArgs>,
-            gesture_fn: &WidgetFn<KeyGestureArgs>,
+            separator_fn: &WidgetFn<KeyGestureSeparatorFnArgs>,
+            modifier_fn: &WidgetFn<ModifierFnArgs>,
+            key_fn: &WidgetFn<KeyFnArgs>,
+            gesture_fn: &WidgetFn<KeyGestureFnArgs>,
         ) {
             let mut gesture_items = ui_vec![];
 
             macro_rules! gen_modifier {
                 ($has:ident, $Variant:ident) => {
                     if gesture.modifiers.$has()
-                        && let Some(n) = modifier_fn.call_checked(ModifierArgs {
-                            gesture: ModifierGesture::$Variant,
+                        && let Some(n) = modifier_fn.call_checked(ModifierFnArgs {
+                            modifier: ModifierGesture::$Variant,
                             is_standalone: false,
                         })
                     {
                         if !gesture_items.is_empty()
-                            && let Some(s) = separator_fn.call_checked(KeyGestureSeparatorArgs {
+                            && let Some(s) = separator_fn.call_checked(KeyGestureSeparatorFnArgs {
                                 between_modifiers: true,
                             })
                         {
@@ -450,9 +495,9 @@ fn generate(shortcut: Shortcuts) -> UiNode {
             gen_modifier!(has_shift, Shift);
             gen_modifier!(has_alt, Alt);
 
-            if let Some(n) = key_fn.call_checked(KeyArgs { key: gesture.key.clone() }) {
+            if let Some(n) = key_fn.call_checked(KeyFnArgs { key: gesture.key.clone() }) {
                 if !gesture_items.is_empty()
-                    && let Some(s) = separator_fn.call_checked(KeyGestureSeparatorArgs { between_modifiers: false })
+                    && let Some(s) = separator_fn.call_checked(KeyGestureSeparatorFnArgs { between_modifiers: false })
                 {
                     gesture_items.push(s);
                 }
@@ -462,7 +507,7 @@ fn generate(shortcut: Shortcuts) -> UiNode {
             if gesture_fn.is_nil() {
                 out.append(&mut gesture_items);
             } else {
-                let gesture = gesture_fn.call(KeyGestureArgs {
+                let gesture = gesture_fn.call(KeyGestureFnArgs {
                     items: gesture_items,
                     gesture,
                 });
@@ -483,7 +528,7 @@ fn generate(shortcut: Shortcuts) -> UiNode {
                     &gesture_fn,
                 );
                 if !shortcut_items.is_empty()
-                    && let Some(s) = chord_separator_fn.call_checked(ChordSeparatorArgs {})
+                    && let Some(s) = chord_separator_fn.call_checked(ChordSeparatorFnArgs {})
                 {
                     shortcut_items.push(s);
                 }
@@ -497,8 +542,8 @@ fn generate(shortcut: Shortcuts) -> UiNode {
                 );
             }
             Shortcut::Modifier(g) => {
-                if let Some(m) = modifier_fn.call_checked(ModifierArgs {
-                    gesture: *g,
+                if let Some(m) = modifier_fn.call_checked(ModifierFnArgs {
+                    modifier: *g,
                     is_standalone: true,
                 }) {
                     shortcut_items.push(m);
@@ -508,7 +553,7 @@ fn generate(shortcut: Shortcuts) -> UiNode {
         if shortcut_fn.is_nil() {
             items.append(&mut shortcut_items);
         } else {
-            let mut s = shortcut_fn.call(ShortcutArgs {
+            let mut s = shortcut_fn.call(ShortcutFnArgs {
                 items: shortcut_items,
                 shortcut: shortcut.clone(),
             });
@@ -520,8 +565,18 @@ fn generate(shortcut: Shortcuts) -> UiNode {
         }
     }
 
-    panel_fn.call(PanelArgs {
+    let mut is_none = false;
+    if items.is_empty() {
+        let none_fn = NONE_FN_VAR.get();
+        if let Some(n) = none_fn.call_checked(NoneFnArgs {}) {
+            items.push(n);
+            is_none = true;
+        }
+    }
+
+    panel_fn.call(PanelFnArgs {
         items,
+        is_none,
         shortcuts: shortcut,
     })
 }